ci: add all test suites to suitespec (#6459)
This change updates `tests/.suitespec.json` to include file listings for
all test suites. This makes CircleCI only run a given test suite when it
can't be certain that no related files changed in a given pull request.
The method I used to find dependencies between contribs was `grep -R
contrib ddtrace/contrib | grep import`.

It also makes a change that finds the pull request number via a
different environment variable as a fallback.

There's a risk with this change that I missed something. If that's the
case there may be some test suites that don't run in cases where they
should. The main reason this is an acceptable risk is that running the
entire test suite on 1.x means that oversights from pull requests will
be caught. Merging a pull request that breaks the 1.x build isn't ideal,
but it's unlikely enough and easy enough to fix by editing
`.suitespec.json` that I believe it's worth the risk.

---------

Co-authored-by: Gabriele N. Tornetta <P403n1x87@users.noreply.github.com>