ci: improve suitespec declarations
This change improves the suitespec component declaration. As a rule
of thumb, patterns matching test sources should not be added to
components to avoid running jobs that depend on source but not on
test scripts. Therefore, we moved as many of these patterns from
the components to the suites.

We also increased the suitespec coverage leaving only a bunch of
sources uncovered. We also improved the suitespec Python API by
adding a templating system whereby the suite name can be
substituted in a pattern during resolution.