fix(tracer): fix sampling mechanism tag propagation bug (#6579)
This change fixes a bug causing `_dd.p.*` span tags not to be inherited
by child spans when there is a process boundary between the parent and
child spans. This makes cross-process parenting match the behavior of
parenting within a single process with respect to tag inheritance.

This change updates all snapshot files and test expectations affected by
this change to include expectation of all relevant `_dd.p.*` tags. There
are also two style changes that were required to get the CI pre_check
step to pass (I'm not sure why that isn't appearing on 1.x).