chore(service naming): Add _dd.base_service tag
This change is being done to propagate global service information to the
backend for the purpose of improving the Service Map.  It has no affect
on the tracer beyond adding additional information.

Prior to this change spans did not contain any contextual information
about the global configuration if span.service was set.

After this change, all spans whose service differs from the globally
configured value will contain a `_dd.base_service` tag which contains
the globally configured value.

Note that the global value by default is set to the value of the
environment variable DD_SERVICE