ci: rename lib-injection job name
The lib injection workflow defined a job called test which might
race and conflict with the test job from the CircleCI
configuration. This would case the state of the test suite to be
reported by a GitHub app that is not the expected one, thus
blocking PRs where the CircleCI run terminates quickly.