fix(datastreams): support datastream propagation with raw message delivery
When a subscription is set up between SNS and a consumer (such as SQS),
there is an option to enable "Raw Message Delivery" which changes the
format of delivered data, primary the MessageAttributes and MessageBody
sections.

Prior to this change, datastream context was dropped/ignored on the
receive because the code assumed the format of delivered messages when
Raw Delivery was disabled.

After this change, the patch handles both raw message delivery and
non-raw delivery.