chore(flask): remove span knowledge from flask wsgi middleware (#6418)
This change removes the term "span" from `_FlaskWSGIMiddleware` in the
interest of clarifying the separation of concerns between
`contrib.flask` and `apm`. The idea is that the Contrib knows about
everything specific to Flask, like how to get information out of headers
and how to package `environ` into a `_RequestType`, and the `apm` module
knows how to create and modify spans.

Because `_FlaskWSGIMiddleware` is a subclass of `_DDWSGIMiddlewareBase`,
it inherits tracing-focused nomenclature in its function signatures,
which this change addresses in two ways: replacing "span" function
arguments with "context" arguments that contain spans, and replacing the
term "span" in function names with the term "call".

Replacing "span" with "call" is a weird change and it's not completely
clear to me that it reduces the net complexity of the `contrib`/`apm`
system. The benefit is that `contrib` engineers don't have to think
about tracing at all; the cost is that `apm` engineers have to look at
the term "call" and understand that in some cases it essentially is a
synonym for "span". This might not be perfect, but I see it as a
reasonable real-world test case for this strict separation of concerns
that we'll learn from and adjust accordingly as code changes.

## Checklist

- [x] Change(s) are motivated and described in the PR description.
- [x] Testing strategy is described if automated tests are not included
in the PR.
- [x] Risk is outlined (performance impact, potential for breakage,
maintainability, etc).
- [x] Change is maintainable (easy to change, telemetry, documentation).
- [x] [Library release note
guidelines](https://ddtrace.readthedocs.io/en/stable/releasenotes.html)
are followed. If no release note is required, add label
`changelog/no-changelog`.
- [x] Documentation is included (in-code, generated user docs, [public
corp docs](https://github.com/DataDog/documentation/)).
- [x] Backport labels are set (if
[applicable](https://ddtrace.readthedocs.io/en/latest/contributing.html#backporting))

## Reviewer Checklist

- [ ] Title is accurate.
- [ ] No unnecessary changes are introduced.
- [ ] Description motivates each change.
- [ ] Avoids breaking
[API](https://ddtrace.readthedocs.io/en/stable/versioning.html#interfaces)
changes unless absolutely necessary.
- [ ] Testing strategy adequately addresses listed risk(s).
- [ ] Change is maintainable (easy to change, telemetry, documentation).
- [ ] Release note makes sense to a user of the library.
- [ ] Reviewer has explicitly acknowledged and discussed the performance
implications of this PR as reported in the benchmarks PR comment.
- [ ] Backport labels are set in a manner that is consistent with the
[release branch maintenance
policy](https://ddtrace.readthedocs.io/en/latest/contributing.html#backporting)

---------

Co-authored-by: Alberto Vara <alberto.vara@datadoghq.com>
Co-authored-by: Brett Langdon <brett.langdon@datadoghq.com>