fix(lib-injection): handle unsupported Python versions
When the injection is applied to an unsupported Python interpreter the
following error is received:

```
ERROR:datadog.autoinstrumentation(pid: 28271): failed to load ddtrace module: No module named 'ddtrace'
Error in sitecustomize; set PYTHONVERBOSE for traceback:
ModuleNotFoundError: No module named 'ddtrace'
```

The application still runs but this adds implementation-detail noise to
the application logs.

A regression test is added for an unsupported Python interpreter just to
ensure that the application still runs.