fix(openai): ensure streamed spans with error are manually finished (#6891)
Resolves #6769.

This PR fixes an unhandled case in our OpenAI integration specifically
regarding streamed `Chat/Completion` requests that have errors and
result in an empty response. In this case, streamed spans were not
finished (as we avoided finishing streamed response spans until the
underlying generator was exhausted) because we did not handle empty
response cases. The fix here is to add manual span finishing for
streamed response spans with error, as we already do that for
non-streamed spans.

The only risk is if there are requests that we trace that might result
in a non-empty response even with an error. This is highly unlikely as
we shouldn't get any response if the corresponding request was faulty.

Note: this PR also moves tagging prompt token usage information to the
`EndpointHook.process_response()` instead of the
`EndpointHook.process_request()` handler as keeping that in the latter
results in prompt token information being recorded, even if there were
no actual prompt/completion operation happening for OpenAI due to the
erroneous request.


## Checklist

- [x] Change(s) are motivated and described in the PR description.
- [x] Testing strategy is described if automated tests are not included
in the PR.
- [x] Risk is outlined (performance impact, potential for breakage,
maintainability, etc).
- [x] Change is maintainable (easy to change, telemetry, documentation).
- [x] [Library release note
guidelines](https://ddtrace.readthedocs.io/en/stable/releasenotes.html)
are followed. If no release note is required, add label
`changelog/no-changelog`.
- [x] Documentation is included (in-code, generated user docs, [public
corp docs](https://github.com/DataDog/documentation/)).
- [x] Backport labels are set (if
[applicable](https://ddtrace.readthedocs.io/en/latest/contributing.html#backporting))

## Reviewer Checklist

- [x] Title is accurate.
- [x] No unnecessary changes are introduced.
- [x] Description motivates each change.
- [x] Avoids breaking
[API](https://ddtrace.readthedocs.io/en/stable/versioning.html#interfaces)
changes unless absolutely necessary.
- [x] Testing strategy adequately addresses listed risk(s).
- [x] Change is maintainable (easy to change, telemetry, documentation).
- [x] Release note makes sense to a user of the library.
- [x] Reviewer has explicitly acknowledged and discussed the performance
implications of this PR as reported in the benchmarks PR comment.
- [x] Backport labels are set in a manner that is consistent with the
[release branch maintenance
policy](https://ddtrace.readthedocs.io/en/latest/contributing.html#backporting)
- [x] If this PR touches code that signs or publishes builds or
packages, or handles credentials of any kind, I've requested a review
from `@DataDog/security-design-and-guidance`.
- [x] This PR doesn't touch any of that.

Co-authored-by: Munir Abdinur <munir.abdinur@datadoghq.com>