chore(dbm): dbm should use peer.service value when enabled
DBM adds comments to database commands for the purposes of context
propagation.  One of the pieces of information in the comments is the
service name of the external database and is used for generating metrics
(i.e. # of writes per second to external database).

Prior to this change, DBM relied on the service name being set to the
value of span.service.  In auto-instrumented services, this value
typically belongs to the fake span generated with the destination
database name, such as "postgres".  This caused issues when peer.service
is enabled, since the fake span no longer existed.

After this change, DBM will use the precursor for peer.service (if
available) when peer service defaults are enabled.