fix(lib-injection): log to syslog rather than stderr
The stderr logs are noisy and not actionable when running many Python
scripts on a system. It is also possible that the injection breaks
scripts and applications that rely on specific stderr output.

Instead, the logs can be written to the system logs. The logs can still
be printed to stderr when using `DD_TRACE_DEBUG=1`.