fix(signals): only handle signals if in main thread
Prior to this change ddtrace attempted to register signal handling code
regardless of whether or not it was in the main thread.  This caused
ddtrace to fail and exit because in Python signal handling can only
happen in the main thread.

After this change, we only add signal handling capabilities if ddtrace
is imported/initialized as part of the main thread.