fix(ci_visibility): fix distributed tracing when using CIVisibility (#7210)
`CIContextProvider` was introduced in #6548 to enable running tests in
this
repo with the `pytest` plugin, but disabled distributed tracing.

`_DD_CIVISIBILITY_USE_CI_CONTEXT_PROVIDER` is introduced as a private
env var that can be used to activate the `CIContextProvider`. It is
false by
default, returning the behavior of the tracer for regular users to what
it was
before the CI context provider was introduced.

This is primarily a band-aid patch. Using the pytest plugin in this repo
remains a WIP.

This change also sets `_DD_CIVISIBILITY_USE_CI_CONTEXT_PROVIDER`
in `riotfile.py` so that the change is transparent in this repo's CI
runs.

## Checklist

- [x] Change(s) are motivated and described in the PR description.
- [x] Testing strategy is described if automated tests are not included
in the PR.
- [x] Risk is outlined (performance impact, potential for breakage,
maintainability, etc).
- [x] Change is maintainable (easy to change, telemetry, documentation).
- [x] [Library release note
guidelines](https://ddtrace.readthedocs.io/en/stable/releasenotes.html)
are followed. If no release note is required, add label
`changelog/no-changelog`.
- [x] Documentation is included (in-code, generated user docs, [public
corp docs](https://github.com/DataDog/documentation/)).
- [x] Backport labels are set (if
[applicable](https://ddtrace.readthedocs.io/en/latest/contributing.html#backporting))

## Reviewer Checklist

- [x] Title is accurate.
- [x] No unnecessary changes are introduced.
- [x] Description motivates each change.
- [x] Avoids breaking
[API](https://ddtrace.readthedocs.io/en/stable/versioning.html#interfaces)
changes unless absolutely necessary.
- [x] Testing strategy adequately addresses listed risk(s).
- [x] Change is maintainable (easy to change, telemetry, documentation).
- [x] Release note makes sense to a user of the library.
- [x] Reviewer has explicitly acknowledged and discussed the performance
implications of this PR as reported in the benchmarks PR comment.
- [x] Backport labels are set in a manner that is consistent with the
[release branch maintenance
policy](https://ddtrace.readthedocs.io/en/latest/contributing.html#backporting)
- [x] If this PR touches code that signs or publishes builds or
packages, or handles credentials of any kind, I've requested a review
from `@DataDog/security-design-and-guidance`.
- [x] This PR doesn't touch any of that.