fix(lib-inject): set correct permissions on package files (#7239)
Package files should not be world-writable. Users noticed injected
ddtrace package files being flagged by security scanners.

In particular the `bytecode` package in Python 3.7 includes
world-writable sources files.

The existing tests should cover that the feature still works in the K8s
library injection. The existing assertion for the permissions on the
files is updated to assert that the permission is correctly applied to
all package files.

For the deb/rpm I manually tested the commands that the build runs and
validated that the files are not world-writable.

There may be unknown situations where Python packages require these
permissions. In the worst-case with these situations the application may
crash due to an unexpected exception at runtime.