chore(ci): pin the library version being built (#7255)
There are no code changes in `1.20.5` only packaging updates so we can
safely use the previous tracer version. This allows us to rapidly get
fix `1.20.5` out without waiting for the corresponding PyPI build to
finish which takes 4+ hours.

The risk is that users who are confused when manually specifying version
`1.20.5` in their K8s chart or deb/rpm install will see ``ddtrace``
version [`1.20.4`](https://pypi.org/project/ddtrace/1.20.4/) installed
instead. There is no harm as the 1.20.5 release is only fixing a
packaging issue for the library injection packages.

This change will be reverted as soon as `1.20.5` is released to ensure
subsequent `1.20` releases include the code changes to the library.