fix(tracing): fix msgpack import error (#7263)
## Description

Reverts: DataDog/dd-trace-py#7099

- This change adds a performance regression to v0.5 and is no longer
required.
- This change fixes an import error raised when `DD_TRACE_API` is set to
`v0.5`. This occurs because msgpack is not installed by ddtrace.

## Motivation
[Import
msgpack](https://github.com/DataDog/dd-trace-py/blob/v2.1.1/ddtrace/internal/_encoding.pyx#L425)
is problematic and breaks v0.5 encoding. To fix this issue we can ensure
the ddtrace library installs msgpack (or we can vendor it) but this
comes with it's own drawbacks. Instead of fixing the importerror this PR
simply reverts the commit that introduced this import.

- [x] checkbox