fix(internal): do not replace sys.modules [backport #6925 to 1.19] (#7070)
Backport of #6925 to 1.19

We refactor the ModuleWatchdog class to avoid replacing the sys.modules
dictionary with a proxy. Instead, we use a WeakValueDictionary to
prevent adding strong references to modules that might be unloaded, and
also preventing the origin mapping from growing unchecked over time.

Fixes #6923.

## Testing Strategy

The existing test suites should be good enough to catch any regressions
from this substantial change. As for the mentioned issue, the change was
tested manually on the reproducer provided in the description provided
by the reporter.

## Risk

This is a substantial change in behaviour for `ModuleWatchdog`. However,
we expect things to run smoother now that `sys.modules` is not _hacked_,
at the cost of a potentially slightly delayed memory clean-up.

## Checklist

- [x] Change(s) are motivated and described in the PR description.
- [x] Testing strategy is described if automated tests are not included
in the PR.
- [x] Risk is outlined (performance impact, potential for breakage,
maintainability, etc).
- [x] Change is maintainable (easy to change, telemetry, documentation).
- [x] [Library release note
guidelines](https://ddtrace.readthedocs.io/en/stable/releasenotes.html)
are followed. If no release note is required, add label
`changelog/no-changelog`.
- [x] Documentation is included (in-code, generated user docs, [public
corp docs](https://github.com/DataDog/documentation/)).
- [x] Backport labels are set (if
[applicable](https://ddtrace.readthedocs.io/en/latest/contributing.html#backporting))

## Reviewer Checklist

- [x] Title is accurate.
- [x] No unnecessary changes are introduced.
- [x] Description motivates each change.
- [x] Avoids breaking
[API](https://ddtrace.readthedocs.io/en/stable/versioning.html#interfaces)
changes unless absolutely necessary.
- [x] Testing strategy adequately addresses listed risk(s).
- [x] Change is maintainable (easy to change, telemetry, documentation).
- [x] Release note makes sense to a user of the library.
- [x] Reviewer has explicitly acknowledged and discussed the performance
implications of this PR as reported in the benchmarks PR comment.
- [x] Backport labels are set in a manner that is consistent with the
[release branch maintenance
policy](https://ddtrace.readthedocs.io/en/latest/contributing.html#backporting)
- [x] If this PR touches code that signs or publishes builds or
packages, or handles credentials of any kind, I've requested a review
from `@DataDog/security-design-and-guidance`.
- [x] This PR doesn't touch any of that.